refactor: search component export for reuse in story
